pyproject.toml
[tool.poetry]
name = "text-generation"
version = "0.1.0"
description = ""
authors = ["Kohei_Ohno <discus0434@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.3"
numpy = "^1.23.1"
beautifulsoup4 = "^4.11.1"
tweepy = "^4.10.0"
python-dotenv = "^0.20.0"
tqdm = "^4.64.0"
jaconv = "^0.3"
tensorflow-macos = "^2.9.2"
tensorflow-metal = "^0.5.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
